Revert to JavaScript
