+++ www/src/components/Plural.js
@@ -14,7 +14,7 @@ import RepositoryArtifacts from './repository/RepositoryArtifacts'
 import RepositoryDeployments from './repository/RepositoryDeployments'
 import RepositoryDescription from './repository/RepositoryDescription'
 import RepositoryEdit from './repository/RepositoryEdit'
-import RepositoryPackages from './repository/RepositoryPackages.tsx'
+import RepositoryPackages from './repository/RepositoryPackages'
 import RepositoryPackagesDocker from './repository/RepositoryPackagesDocker'
 import RepositoryPackagesHelm from './repository/RepositoryPackagesHelm'
 import RepositoryPackagesTerraform from './repository/RepositoryPackagesTerraform'
