+++ www/src/components/repository/RepositoryPackages.js
@@ -1,0 +1,71 @@
+import { Outlet, useLocation, useNavigate } from 'react-router-dom'
+import { Flex } from 'honorable'
+
+import { useState } from 'react'
+import { Input, MagnifyingGlassIcon } from 'pluralsh-design-system'
+
+import { ButtonGroup } from '../utils/ButtonGroup'
+
+export function packageCardStyle(first, last) {
+  return {
+    backgroundColor: 'fill-one',
+    hoverIndicator: 'fill-two',
+    color: 'text',
+    textDecoration: 'none',
+    border: '1px solid border-fill-two',
+    borderTop: first ? '1px solid border-fill-two' : 'none',
+    borderTopLeftRadius: first ? '4px' : 0,
+    borderTopRightRadius: first ? '4px' : 0,
+    borderBottomLeftRadius: last ? '4px' : 0,
+    borderBottomRightRadius: last ? '4px' : 0,
+    align: 'center',
+    px: 1,
+    py: 0.5,
+  }
+}
+const tabToUrl = {
+  'Helm Charts': 'helm',
+  'Terraform Modules': 'terraform',
+  'Docker Repositories': 'docker',
+}
+
+export default function RepositoryPackages() {
+  const [query, setQuery] = useState('')
+  const { pathname } = useLocation()
+  const navigate = useNavigate()
+  const tabUrl = pathname.substring(pathname.lastIndexOf('/') + 1)
+
+  return (
+    <Flex
+      direction="column"
+      height="100%"
+    >
+      <Flex justifyContent="space-between">
+        <Input
+          flexBasis="350px"
+          marginRight="medium"
+          startIcon={(
+            <MagnifyingGlassIcon
+              size={14}
+            />
+          )}
+          placeholder="Search a package"
+          value={query}
+          onChange={event => setQuery(event.target.value)}
+        />
+        <ButtonGroup
+          tabs={Object.keys(tabToUrl)}
+          default={Object.keys(tabToUrl).find(tab => tabToUrl[tab] === tabUrl) || 'Helm Charts'}
+          onChange={tab => navigate(tabToUrl[tab])}
+        />
+      </Flex>
+      <Flex
+        mt={1}
+        direction="column"
+        flexGrow={1}
+      >
+        <Outlet context={[query, setQuery]} />
+      </Flex>
+    </Flex>
+  )
+}
