+++ www/src/components/utils/hooks.js
@@ -10,15 +10,3 @@ import { withMentions } from './TypeaheadEditor'
 export function useEditor() {
   return useMemo(() => withMentions(withReact(withHistory(createEditor()))), [])
 }
-
-export function withRouter(Component) {
-  return function (props) {
-    return (
-      <Component
-        {...props}
-        navigate={useNavigate()}
-        location={useLocation()}
-      />
-    )
-  }  
-}
